requirements.txt
# Core dependencies
sqlalchemy>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# CLI and output
typer>=0.9.0
rich>=13.0.0

# Reliability
tenacity>=8.2.0

# Validation
pydantic>=2.0.0

# Caching
diskcache>=5.6.0

# Database
alembic>=1.12.0

# Optional: API clients
beautifulsoup4>=4.12.0
lxml>=4.9.0
